requirements.txt
absl-py==0.8.0
alpha-vantage==2.1.1
anaconda==0.0.1.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==1.5.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==4.0.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.4.1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.2.1
matplotlib==3.1.2
mistune==0.8.4
mysql-connector==2.2.9
nbconvert==5.6.0
nbformat==4.4.0
newsapi-python==0.2.5
notebook==6.0.1
numpy==1.17.2
opencv-python==4.1.2.30
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
polyglot==16.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
PuLP==1.6.10
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.4
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.21.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
simpy==3.0.11
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
SQLAlchemy==1.3.10
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.3
virtualenv==16.7.8
wcwidth==0.1.7
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
